some improvements for corner cases:
- do not allow freeze if shadow dir is not empty

Current behaviour leads to some sophisticated cases during restore: if
shadow dir already had any partitions, those are either old, because
partitions were overwritten, or same and a new hard link will be
created. In both of these cases files will be uploaded to S3. During
restore there is no method to distinguish such parts.
So, we need to have shadow dir empty, to have consistent current copy
of data in Clickhouse. This can be easily done via clean command.

- do not throw error if there are no partitions in Clickhouse

In case there are no actual data (partitions) in Clickhouse just log
a message about this and quit.

- do not throw error during restore if there are no partitions

In case a backup was made which does not have partitions (shadow dir is
empty) but metadata is there and may be used for creating tables, we
just log a message, that there is nothing to do.

- minor change in create-tables when getting Clickhouse data path