server: Add callbacks query param
Add callbacks query param, which can be used by clients to the
server for notifications on completion of an asynchronous operation.

Callback support is added to the following endpoints:

    /backup/create
    /backup/upload/{name}
    /backup/download/{name}
    /backup/restore/{name}

The callback parameter can be passed to the target endpoint with
the addition of

?callback={URI Encoded Callback URL}

Callback payloads are POSTed with the JSON serialized struct
CallbackResponse, which is exported for use by clients. Multiple
callback URLs can be specified; all URLs will be hit regardless of
failures in any single endpoint.