backup: Return nil error for delete of nonexistent directory
Make it so that deleting a directory is idempotent, returning a nil
error if the directory is already deleted. This makes it so that
one can call the Clean function in a way that it ensures the local
state is clean rather than needing to be concerned about whether or
not the shadow directory is existent at the time.

Fixes #686