gcs: mitigate transient errors causing corruption
For long backups, there appears to be a relatively high chance
of random issue like a 503. When these issues, it appears that
the clickhouse-backup retries are not sufficient, as we are seeing
missing table parts get reported.

This is a mitigation because in theory, the wrapping retries would
correctly retry these issues, however it does not appear to be the
case. More understanding is needed to fix that particular issue.